refactor: remove gameInfo from Card, move to component
