+++ src/components/Card/Card.module.scss
@@ -21,7 +21,6 @@
     align-items: center;
     justify-content: center;
     padding-top: 1rem;
-    // width: 200px;
     flex: 1;
   }
 
@@ -40,27 +39,5 @@
     display: flex;
     align-items: center;
     justify-content: center;
-  }
-  .info {
-    font-size: 0.9rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-    .info_item {
-      display: flex;
-      align-items: center;
-
-      i {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-
-        width: 2rem;
-      }
-      .item_data {
-        margin-left: 0.2rem;
-        font-weight: bold;
-      }
-    }
   }
 }
