+++ src/components/Card/Card.tsx
@@ -4,6 +4,8 @@ import React, { Component, ReactNode } from 'react';
 
 import { Link } from 'react-router-dom';
 
+import GameInfo from '../GameInfo';
+
 import styles from './Card.module.scss';
 
 const PATH_TO_GAME_IMG = 'assets/images/games_webp/';
@@ -32,34 +34,6 @@ class Card extends Component<ICardProps> {
       publisherTitle,
     } = this.props;
 
-    const infoData = [
-      {
-        icon: 'fa-solid fa-person',
-        title: 'Кол-во игроков:',
-        title_info: players,
-      },
-      {
-        icon: 'fa-solid fa-clock',
-        title: 'Время партии:',
-        title_info: playingTime,
-      },
-      {
-        icon: 'fa-solid fa-child-reaching',
-        title: 'Возраст:',
-        title_info: `от ${age} лет`,
-      },
-      {
-        icon: 'fa-solid fa-language',
-        title: 'Язык:',
-        title_info: lang,
-      },
-      {
-        icon: 'fa-solid fa-calendar-days',
-        title: 'Дата релиза:',
-        title_info: releaseDate,
-      },
-    ];
-
     return (
       <li className={styles.card}>
         <div className={styles.image}>
@@ -71,25 +45,14 @@ class Card extends Component<ICardProps> {
         <div>{isGame ? 'Игра' : 'Дполнение к игре'}</div>
 
         <div className={styles.rating}>
-          BGG :<span>{scoreBGG}</span>
-          <pre> | </pre>Tesera <span>{scoreTesera}</span>
+          BGG: <span>{scoreBGG}</span>
+          <pre> | </pre>Tesera: <span>{scoreTesera}</span>
         </div>
 
         <div className={styles.publisher}>
           <img src={PATH_TO_PUBLISHER_IMG + publisherImage} alt={publisherTitle} />
         </div>
-
-        <div className={styles.info}>
-          {infoData.map((infoItem) => {
-            const { icon, title, title_info } = infoItem;
-            return (
-              <span className={styles.info_item} key={icon}>
-                <i className={icon}></i> {title}
-                <span className={styles.item_data}> {title_info}</span>
-              </span>
-            );
-          })}
-        </div>
+        <GameInfo {...{ releaseDate, players, age, playingTime, lang }} />
       </li>
     );
   }
